+++ benchmark/Benchmarks.hs
@@ -30,6 +30,7 @@ main = defaultMain
         [ bench "size"  $ nf TS.size bigSeries
         , bench "valueAt"  $ nf (\xs -> TS.valueAt xs lastIndex) bigSeries
         , bench "slice"  $ nf (\xs -> TS.valueAt (TS.slice xs (index 2) lastIndex) (index 1)) bigSeries
+        , bench "max"  $ nf maximum bigSeries
         , bench "fmap"  $ nf (fmap (+ 2)) bigSeries
         ]
     , bgroup "IO"