+++ test-src/Data/TimeSeries/SeriesSpec.hs
@@ -53,9 +53,9 @@ spec = do
 
     it "maximum value" $ do
         let xs = TS.tsSeries [1..] [10.0, 1.2, 32.4, 0.65, 11.0]
-        TS.max xs `shouldBe` 32.4
+        maximum xs `shouldBe` 32.4
 
     it "minimum value" $ do
         let xs = TS.tsSeries [1..5] [10.0, 1.2, 32.4, 0.65, 11.0]
-        TS.min xs `shouldBe` 0.65
+        minimum xs `shouldBe` 0.65